Move compute_raw from _dev_repo_metrics.py to _raw_metrics.py
Co-locate RawMetrics dataclass with its factory function.
_dev_repo_metrics.py drops from 49 to 24 lines — now only
wraps compute_raw with grouping logic (by dev, by repo).
