+++ git_dev_metrics/metrics/_dev_repo_metrics.py
@@ -1,31 +1,6 @@
 from ..models import PullRequest
-from ._raw_metrics import RawMetrics
-from .calculator import (
-    calculate_ai_percentage,
-    calculate_avg_lines_per_pr,
-    calculate_cycle_time,
-    calculate_pickup_time,
-    calculate_pr_size,
-    calculate_prs_per_week,
-    calculate_review_time,
-    calculate_reviews_given,
-    calculate_throughput,
-    group_prs_by_devs,
-)
-
-
-def compute_raw(prs: list[PullRequest], days: int, reviews_given: int) -> RawMetrics:
-    return RawMetrics(
-        cycle_time=calculate_cycle_time(prs),
-        pr_size=calculate_pr_size(prs),
-        avg_lines_per_pr=calculate_avg_lines_per_pr(prs),
-        pr_count=calculate_throughput(prs),
-        pickup_time=calculate_pickup_time(prs),
-        review_time=calculate_review_time(prs),
-        prs_per_week=calculate_prs_per_week(prs, days),
-        reviews_given=reviews_given,
-        ai_percentage=calculate_ai_percentage(prs),
-    )
+from ._raw_metrics import RawMetrics, compute_raw
+from .calculator import calculate_reviews_given, group_prs_by_devs
 
 
 def compute_dev_metrics(
